backend/requirements.txt
azure-ai-inference
azure-core
colorlog
dotenv
python-dotenv
requests
fastapi
pydantic
uvicorn
langchain
langchain-community
langchain-openai
markdown
GitPython